[REFACTOR] Organise admin view styles
