--- src/client/react/admin/scss/_team-profile.scss
@@ -1,35 +1,0 @@
-#team-profile {
-	.member {
-		max-width: 18rem;
-		margin: 0.3rem 0;
-		display: flex;
-		flex-wrap: wrap;
-		justify-content: space-between;
-		align-items: center;
-		background: #f0f0f0;
-		border-radius: 0.5rem;
-		padding: 0.5rem;
-	}
-
-	.action-button {
-		margin-top: -0.3rem;
-		margin-left: 0.5rem;
-	}
-
-	.manage {
-		display: flex;
-		flex-wrap: wrap;
-		flex-direction: column;
-		justify-content: space-between;
-
-		.manage-item {
-			display: flex;
-			align-items: center;
-			margin-bottom: 0.3rem;
-		}
-	}
-}
-
-.pt-dialog .pt-callout {
-	margin-bottom: 0.5rem;
-}
