+++ src/client/react/admin/scss/components/_navbar.scss
