+++ src/client/react/admin/scss/pages/_dashboard.scss
@@ -1,4 +1,4 @@
-@import '../../scss/variables/bp';
+@import '../../../scss/variables/bp';
 
 #admin-dashboard {
 	@media (min-width: $bp-s) {
