+++ src/client/react/admin/scss/views/_gamestate-view.scss
