+++ src/client/react/admin/scss/views/_users-view.scss
@@ -1,4 +1,4 @@
-@import '../../scss/variables/bp';
+@import '../../../scss/variables/bp';
 
 #dashboard-users {
 	.user-card {
